backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
google-genai==1.0.0
gunicorn==23.0.0
python-dotenv==1.0.1